+++ src/functions/model/run-code-generator-test-object.spec.ts
@@ -131,16 +131,28 @@ describe('ModelRunCodeGeneratorForTypeScriptTestObject', () => {
     ).toThrow(NoImplementationFoundError);
   });
 
+  it('should throw an error if the input data cannot be created', async () => {
+    const { context } = createContext({
+      projectPath: tmpDir,
+      configuration: baseConfiguration,
+      // ModelMakeGeneratorQuicktypeInputData will not be present.
+      functions: [ModelRunCodeGeneratorForTypeScriptTestObject],
+    });
+
+    const actualPromise = context.call(ModelRunCodeGenerator, baseArguments);
+
+    await expect(actualPromise).rejects.toThrow(
+      'Could not generate input data for code generation. Ensure the model schema format is supported.',
+    );
+  });
+
   it('should generate TypeScript test object functions from JSON schemas and return GeneratedSchemas', async () => {
     const schemaFile = join(tmpDir, 'person.schema.json');
     const modelFile = join(tmpDir, 'model.ts');
     await writeFile(schemaFile, JSON.stringify(SCHEMA));
     const { context, functionRegistry } = createContext({
       projectPath: tmpDir,
-      configuration: {
-        ...baseConfiguration,
-        project: { ...baseConfiguration.project!, language: 'typescript' },
-      },
+      configuration: baseConfiguration,
       functions: [ModelRunCodeGeneratorForTypeScriptTestObject],
     });
     const input = await makeJsonSchemaInputData([schemaFile]);